fix: put access control 404 when service provider not found
Setting access control (via PUT access endpoint) responds 404 when the Service Provider object is not found so that the internal metadata database does not get corrupted with orphaned policies. Also amended the endpoint's method from POST to PUT since I believe that is more proper because of the endpoint's idempotency.